Update request_builder.go
Change the OrderBy function parameter to be `string, bool` instead of `string, string`. It's a trivial change but since Order can only either be asc or desc, I think it makes more sense to be a boolean value thing having to specify which. 

And because this isn't in the tag/ released version it won't be a breaking change